requirements.txt
mysql-connector==2.2.9
numpy==2.3.5
pandas==2.3.3
progress==1.6.1
progress_bar==8
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
